requirements.txt
fastapi==
uvicorn==
python-dotenv==
pydantic==
requests==
beautifulsoup4==
arxiv==
openai==
tenacity==
python-multipart==
psutil==
